rename model to be more specific
