+++ tests/unit/test_prepare_non_res_ind_cqc_features_ascwds_inc_dormancy.py
@@ -2,15 +2,15 @@ import unittest
 import warnings
 from unittest.mock import ANY, Mock, patch
 
-import jobs.prepare_non_res_ind_cqc_features as job
+import jobs.prepare_non_res_ind_cqc_features_ascwds_inc_dormancy as job
 from utils import utils
 
 from utils.column_names.ind_cqc_pipeline_columns import (
     PartitionKeys as Keys,
 )
 
-from tests.test_file_data import NonResFeaturesData as Data
-from tests.test_file_schemas import NonResFeaturesSchema as Schemas
+from tests.test_file_data import NonResAscwdsWithDormancyFeaturesData as Data
+from tests.test_file_schemas import NonResAscwdsWithDormancyFeaturesSchema as Schemas
 
 
 class NonResLocationsFeatureEngineeringTests(unittest.TestCase):
